Change int to long long

+++ holberton.h
@@ -5,7 +5,7 @@
 int _printf(const char *format, ...);
 int _putchar(char c);
 void print_unum(unsigned int n);
-void print_number(int n);
+void print_number(long long int n);
 void rev_string(char *);
 unsigned int _convert(char a, va_list args, unsigned int chars);
 unsigned int _puts(char *str, unsigned int chars);
